pyproject.toml
[project]
name = "clepsy-desktop-source"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncclick>=8.1.7.2",
    "fasteners>=0.19",
    "httpx>=0.28.0",
    "loguru>=0.7.3",
    "mss>=10.0.0",
    "pillow>=11.0.0",
    "platformdirs>=4.3.8",
    "pydantic-settings>=2.6.1",
    "pynput>=1.7.7 ; sys_platform == 'win32' or sys_platform == 'darwin'",
    "pyside6==6.9.2",
    "pystray>=0.19.5",
    "pywin32>=310 ; sys_platform == 'win32'",
    "pywinctl>=0.4.1",
    # Linux-specific backends for idle detection
    "dbus-next>=0.2.3 ; sys_platform == 'linux'",
    "python-xlib>=0.33 ; sys_platform == 'linux'",
    "pyqtdarktheme>=2.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/clepsy_desktop_source"]

[tool.uv]
dev-dependencies = [
    "pyinstaller>=6.14.1",
]

python-preference = "managed"

[tool.uv.sources]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.12 (matches project requires-python)
target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.pylint."MESSAGES CONTROL"]
disable = [
    "logging-fstring-interpolation", 
   "attribute-defined-outside-init"
]
